address review

+++ src/smt/proof_post_processor.cpp
@@ -144,11 +144,11 @@ Node ProofPostprocessCallback::eliminateCrowdingLits(
   // them. The factoring step must be added after the last inclusion and before
   // its elimination.
   std::unordered_set<TNode, TNodeHashFunction> crowding;
-  std::vector<std::pair<Node, unsigned>> lastInclusion;
+  std::vector<std::pair<Node, size_t>> lastInclusion;
   // positions of eliminators of crowding literals, which are the positions of
   // the clauses that eliminate crowding literals *after* their last inclusion
-  std::vector<unsigned> eliminators;
-  for (unsigned i = 0, size = clauseLits.size(); i < size; ++i)
+  std::vector<size_t> eliminators;
+  for (size_t i = 0, size = clauseLits.size(); i < size; ++i)
   {
     if (!crowding.count(clauseLits[i])
         && std::find(
@@ -161,7 +161,7 @@ Node ProofPostprocessCallback::eliminateCrowdingLits(
       // position of the last resolution link that introduces the crowding
       // literal. Note that this position has to be *before* the last link, as a
       // link *after* the last inclusion must eliminate the crowding literal.
-      unsigned j;
+      size_t j;
       for (j = children.size() - 1; j > 0; --j)
       {
         // notice that only non-unit clauses may be introducing the crowding
@@ -217,7 +217,7 @@ Node ProofPostprocessCallback::eliminateCrowdingLits(
   Assert(!lastInclusion.empty());
   // order map so that we process crowding literals in the order of the clauses
   // that last introduce them
-  auto cmp = [=](std::pair<Node, unsigned>& a, std::pair<Node, unsigned>& b) {
+  auto cmp = [=](std::pair<Node, size_t>& a, std::pair<Node, size_t>& b) {
     return a.second < b.second;
   };
   std::sort(lastInclusion.begin(), lastInclusion.end(), cmp);
@@ -226,13 +226,13 @@ Node ProofPostprocessCallback::eliminateCrowdingLits(
   if (Trace.isOn("smt-proof-pp-debug"))
   {
     Trace("smt-proof-pp-debug") << "crowding lits last inclusion:\n";
-    for (const std::pair<const Node&, unsigned>& pair : lastInclusion)
+    for (const std::pair<const Node&, size_t>& pair : lastInclusion)
     {
       Trace("smt-proof-pp-debug")
           << "\t- [" << pair.second << "] : " << pair.first << "\n";
     }
     Trace("smt-proof-pp-debug") << "eliminators:";
-    for (unsigned elim : eliminators)
+    for (size_t elim : eliminators)
     {
       Trace("smt-proof-pp-debug") << " " << elim;
     }
@@ -270,16 +270,16 @@ Node ProofPostprocessCallback::eliminateCrowdingLits(
   // step, the first crowded literal to have its last inclusion after "end". The
   // first elimination position to be bigger than the position of that crowded
   // literal is the next guarded elimination position.
-  unsigned lastElim = 0;
+  size_t lastElim = 0;
   Node lastClause = children[0];
   std::vector<Node> childrenRes;
   std::vector<Node> childrenResArgs;
   Node resPlaceHolder;
-  unsigned nextGuardedElimPos = eliminators[0];
+  size_t nextGuardedElimPos = eliminators[0];
   do
   {
-    unsigned start = lastElim + 1;
-    unsigned end = nextGuardedElimPos - 1;
+    size_t start = lastElim + 1;
+    size_t end = nextGuardedElimPos - 1;
     Trace("smt-proof-pp-debug2")
         << "res with:\n\tlastClause: " << lastClause << "\n\tstart: " << start
         << "\n\tend: " << end << "\n";
@@ -329,8 +329,8 @@ Node ProofPostprocessCallback::eliminateCrowdingLits(
     lastElim = end;
 
     // find the position of the last inclusion of the next crowded literal
-    unsigned nextCrowdedInclusionPos = lastInclusion.size();
-    for (unsigned i = 0, size = lastInclusion.size(); i < size; ++i)
+    size_t nextCrowdedInclusionPos = lastInclusion.size();
+    for (size_t i = 0, size = lastInclusion.size(); i < size; ++i)
     {
       if (lastInclusion[i].second > lastElim)
       {
@@ -351,7 +351,7 @@ Node ProofPostprocessCallback::eliminateCrowdingLits(
     else
     {
       nextGuardedElimPos = children.size();
-      for (unsigned i = 0, size = eliminators.size(); i < size; ++i)
+      for (size_t i = 0, size = eliminators.size(); i < size; ++i)
       {
         //  nextGuardedElimPos is the largest element of
         // eliminators bigger the next crowded literal's last inclusion
@@ -696,7 +696,7 @@ Node ProofPostprocessCallback::expandMacros(PfRule id,
       // different
       std::vector<Node> factoredLits;
       std::unordered_set<TNode, TNodeHashFunction> clauseSet;
-      for (unsigned i = 0, size = chainConclusionLits.size(); i < size; ++i)
+      for (size_t i = 0, size = chainConclusionLits.size(); i < size; ++i)
       {
         if (clauseSet.count(chainConclusionLits[i]))
         {
